# Speed up checksum