Uses a BufReader in the default implementation of `Checksum` trait for Read + Seek + ?Sized. 
Switched from File to the traits (which are all true for File) but could add some flexibility if needed elsewhere.

@xpromache I'd be interested if you want to test this on your 200MB test file. I've only tried on a 5MB file so far.
I've attached the nightly toolchain code I used in a dummy repo with a `/data` folder that housed the `large.txt` test file from this repo.


fixes #12

<Summary>Bench Example</Summary>
<details>

```rust
#![feature(test)]
use std::{
    fs::File,
    io::{BufRead, BufReader, Read, Seek},
    path::Path,
};

fn main() {}

fn new_checksum() -> Result<u32, Box<dyn std::error::Error>> {
    let data = File::open(
        Path::new(env!("CARGO_MANIFEST_DIR"))
            .join("data")
            .join("large.txt"),
    )?;
    let mut reader = BufReader::new(data);
    reader.rewind()?;
    let mut checksum: u32 = 0;
    'outer: loop {
        let buffer = reader.fill_buf()?;
        if buffer.is_empty() {
            break 'outer;
        }

        let mut iter = buffer.chunks_exact(4);
        (&mut iter).for_each(|chunk| {
            checksum = checksum
                .overflowing_add(u32::from_be_bytes(chunk.try_into().unwrap()))
                .0;
        });
        let mut remainder = iter.remainder().to_vec();
        remainder.resize(4, 0_u8);
        checksum = checksum
            .overflowing_add(u32::from_be_bytes(remainder.try_into().unwrap()))
            .0;

        let len = buffer.len();
        reader.consume(len);
    }
    Ok(checksum)
}

fn old_checksum() -> Result<u32, Box<dyn std::error::Error>> {
    let mut reader = File::open(
        Path::new(env!("CARGO_MANIFEST_DIR"))
            .join("data")
            .join("large.txt"),
    )?;
    let mut checksum = 0_u32;
    // reset the file pointer to the beginningouter
    // let mut reader = BufReader::new(Read::by_ref(self));
    reader.rewind()?;

    loop {
        let mut u32_buff = Vec::with_capacity(4);
        // this line will take 4 bytes at a time from the file.
        // if there are less that 4 bytes left it will right pad the buffer with 0s.
        let num_read = Read::take(&mut reader, 4).read_to_end(&mut u32_buff)?;
        match num_read {
            // if nothing was read break from the loop
            0 => break,
            // otherwise add the 4 bytes as a u32 to the checksum
            num => {
                // If the file size is not a multiple of 4
                // right pad the final u32 with 0s.
                if num < 4 {
                    u32_buff.resize(4, 0_u8);
                }
                let val = u32::from_be_bytes([u32_buff[0], u32_buff[1], u32_buff[2], u32_buff[3]]);
                (checksum, _) = checksum.overflowing_add(val);
            }
        }
    }
    Ok(checksum)
}

#[cfg(test)]
mod tests {
    extern crate test;

    use super::*;
    use test::Bencher;

    #[test]
    fn equality() {
        assert_eq!(old_checksum().unwrap(), new_checksum().unwrap())
    }

    #[bench]
    fn old(b: &mut Bencher) {
        b.iter(|| test::black_box(old_checksum()))
    }

    #[bench]
    fn new(b: &mut Bencher) {
        b.iter(|| test::black_box(new_checksum()))
    }
}
```
</details>

Output from benching on the `large.txt` file from this repo

```
-> % cargo +nightly bench
   Compiling rust_playground v0.1.0 (~/src/rust_playground)
    Finished bench [optimized] target(s) in 5.00s
     Running unittests src/main.rs (target/release/deps/rust_playground-2495cce57173af5a)

running 3 tests
test tests::equality ... ignored
test tests::new ... bench:   1,603,494 ns/iter (+/- 161,756)
test tests::old ... bench: 1,347,861,500 ns/iter (+/- 572,146,376)

test result: ok. 0 passed; 0 failed; 1 ignored; 2 measured; 0 filtered out; finished in 399.25s
```