@xpromache 2023-03-02
Works perfectly now! 
Checksum computed in about 75 milliseconds :)